Day 21, part 2 incomplete
Switching to Lagrange interpolation - not correct yet, but close I think

+++ 2023/Day21.h
@@ -7,7 +7,7 @@ private:
 	const std::string testFile2 = "inputs/day21_part2_test.txt";
 	const std::string puzzleFile = "inputs/day21.txt";
 	int calculatePuzzle1(std::vector<std::string> input);
-	int calculatePuzzle2(std::vector<std::string> input);
+	uint64_t calculatePuzzle2(std::vector<std::string> input);
 public:
 	void puzzle1();
 	void puzzle2();
